requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
jsonschema>=4.0.0
pytest>=8.0.0
